requirements.txt
raspend==2.0.3
tzlocal==2.0.0
pyserial==3.4
